将control.cpp中scale的算法进行优化，避免整数除法导致scale == 0致使底盘解算setspeed = 0，出现有speedxy但是setcurrent = 0的情况

+++ STM32F405/control.cpp
@@ -164,7 +164,7 @@ void CONTROL::CHASSIS::Mecanum_Resolve(
 
 	if (max_abs_speed > para.max_speed)
 	{
-		scale = para.max_speed / max_abs_speed;
+		scale = static_cast<float>(para.max_speed) / max_abs_speed;
 	}
 
 	/*
